Enable GHC call-stack simulation in "safe" partial functions.
`(!)` etc. are of course only safer than their `unsafe` pendants in the sense that
errors are raised as proper exceptions, but when used in sizable applications
that by itself does not make the errors easy to find.

GHC-8 has added support for simulated call stack which can actually give some
information on the use site of such errors. In this commit, I added the
necessary constraint to the partial functions which perform bounds checks.